Move feature drawer logic to block based track
