+++ packages/linear-genome-view/src/BasicTrack/blockBasedTrackModel.ts
@@ -1,6 +1,11 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import CompositeMap from '@gmod/jbrowse-core/util/compositeMap'
-import { getContainingView } from '@gmod/jbrowse-core/util/tracks'
+import {
+  getContainingView,
+  getParentRenderProps,
+} from '@gmod/jbrowse-core/util/tracks'
 import { autorun } from 'mobx'
+import { getSession } from '@gmod/jbrowse-core/util'
 import { addDisposer, types, Instance } from 'mobx-state-tree'
 import BlockState, { BlockStateModel } from './util/serverSideRenderedBlock'
 import baseTrack from './baseTrackModel'
@@ -42,7 +47,27 @@ const blockBasedTrack = types.compose(
       get blockDefinitions() {
         return getContainingView(self)[this.blockType]
       },
+      /**
+       * returns a string feature ID if the globally-selected object
+       * is probably a feature
+       */
+      get selectedFeatureId() {
+        const session = getSession(self) as any
+        if (!session) return undefined
+        const { selection } = session
+        // does it quack like a feature?
+        if (
+          selection &&
+          typeof selection.get === 'function' &&
+          typeof selection.id === 'function'
+        ) {
+          // probably is a feature
+          return selection.id()
+        }
+        return undefined
+      },
     }))
+
     .actions(self => ({
       afterAttach() {
         // watch the parent's blocks to update our block state when they change
@@ -77,6 +102,43 @@ const blockBasedTrack = types.compose(
 
       deleteBlock(key: string) {
         self.blockState.delete(key)
+      },
+
+      selectFeature(feature: Feature) {
+        const session = getSession(self) as any
+        // TODO: we shouldn't need to have to get this deep into knowing about
+        // drawer widgets here, the drawer widget should be a reaction to
+        // setting a selected feature...right???
+        if (session.drawerWidgets) {
+          const featureWidget = session.addDrawerWidget(
+            'AlignmentsFeatureDrawerWidget',
+            'alignmentsFeature',
+            // @ts-ignore
+            { featureData: feature.data as any },
+          )
+          session.showDrawerWidget(featureWidget)
+        }
+        session.setSelection(feature)
+      },
+      clearFeatureSelection() {
+        const session = getSession(self) as any
+        session.clearSelection()
+      },
+    }))
+
+    .views(self => ({
+      get composedRenderProps() {
+        return {
+          ...getParentRenderProps(self),
+          trackModel: self,
+          onFeatureClick(event: any, featureId: string) {
+            const feature = self.features.get(featureId)
+            self.selectFeature(feature as Feature)
+          },
+        }
+      },
+      get renderProps() {
+        return this.composedRenderProps
       },
     }))
     .postProcessSnapshot(self => {
